+++ src/features/Class/ClassPage/__test__/columns.test.jsx
@@ -1,6 +1,6 @@
 import '@testing-library/jest-dom/extend-expect';
 
-import { columns } from 'features/Lessons/LessonDetailPage/columns';
+import { columns } from 'features/Class/ClassPage/columns';
 
 describe('columns', () => {
   test('Should return an array of columns with correct properties', () => {
